I made the screen shake
When you get shot or run into an enemy. I also made the backgrounds have Public floats for speed so we can see what works best. The HUD needs to be fixed so I will leave that to Maher still but I'll keep adding sounds. I added an ambiance for the background. Kinda loud rn but we can change that.